weaviate-admin-api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
weaviate-client==3.25.3
PyJWT==2.8.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
email-validator==2.2.0

